Initial Commit
Added temperature correction for pKa values with input for d(pKa)/dt. Added temperature correction curve for autoionization of water. Introduced a new ionic activity / concentration correction function based on an application of Huckel / Davies.

Initial attempt at ionic activity correction improves the accuracy of the model compared with readings taken on an actual buffer when temperature information is provided, but some of the technical details seem suspect. Not sure if ionic activity should be calculated for the entire solution, and a uniform correction should be applied to all pKa values, or if corrections should be calculated and applied on a species-by-species basis.